backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
python-dotenv
supabase
requests
python-jose[cryptography]
pydantic[email]
python-dateutil
backports.zoneinfo
email-validator